+++ test/mocks/ImageAngleAdjusterMock.h
@@ -5,8 +5,8 @@
 
 class ImageAngleAdjusterMock : public d3t12::ImageAngleAdjuster {
 public:
-  MOCK_METHOD1(adjustX, void(int degrees));
-  MOCK_METHOD1(adjustY, void(int degrees));
+  MOCK_METHOD1(adjustX, void(float degrees));
+  MOCK_METHOD1(adjustY, void(float degrees));
 };
 
 #endif
